+++ src/main/de/rwth/dbis/acis/bazaar/service/dal/repositories/ProjectRepository.java
@@ -35,5 +35,5 @@ public interface ProjectRepository extends Repository<Project> {
 
     List<Project> findAllPublic(Pageable pageable) throws BazaarException;
 
-    List<Project> findAllPublicAndAuthorized(PageInfo pageable, int userId) throws BazaarException;
+    List<Project> findAllPublicAndAuthorized(PageInfo pageable, long userId) throws BazaarException;
 }
